added adapter in view

+++ app/src/main/java/com/riteh/autoshare/responses/weather/forecast/Minutely.kt
@@ -2,5 +2,5 @@ package com.riteh.autoshare.responses.weather.forecast
 
 data class Minutely(
     val dt: Int,
-    val precipitation: Int
+    val precipitation: Float
 )
